feat(euler): solve Problems 25-27 with benchmarks and math documentation
- P25: 1000-Digit Fibonacci Number (4782) — big number digit-array addition, 3.67ms
- P26: Reciprocal Cycles (983) — multiplicative order of 10 mod d, 216µs
- P27: Quadratic Primes (-59231) — n²-61n+971 produces 71 consecutive primes, 3.27ms
  Discovered: winning formula is Euler's m²+m+41 reflected and shifted by 30,
  with perfect parabola symmetry and the inescapable 41² barrier

+++ project_euler/src/main.rs
@@ -31,7 +31,7 @@ fn main() {
     });
 
     match ProblemSolver::solve(problem_num) {
-        Some(answer) => println!("Problem {}: {}", problem_num, answer),
+        Some(answer) => println!("Problem {}: {}", problem_num, answer as i64),
         None => {
             println!("Problem {} not yet implemented", problem_num);
             println!("\nImplemented problems: {:?}", ProblemSolver::implemented());
@@ -65,7 +65,7 @@ fn solve_all() {
 
     for &num in &problems {
         if let Some(answer) = ProblemSolver::solve(num) {
-            println!("Problem {:3}: {}", num, answer);
+            println!("Problem {:3}: {}", num, answer as i64);
         }
     }
 }
